requirements.txt
python-telegram-bot==20.3
aiohttp==3.9.0
asyncio==3.4.3
Pillow==10.1.0
APScheduler==3.10.4
python-dotenv==1.0.0
urllib3==2.0.7
PyJWT==2.8.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
